initial round of removes and refactors for legacy balsam executor
