+++ docs/executor/balsam_2_executor.rst
@@ -1,14 +1,16 @@
+.. _balsam-exctr:
+
 Balsam Executor - Remote apps
 =============================
 
-.. automodule:: balsam_executor
+.. automodule:: libensemble.executors.balsam_executor
   :no-undoc-members:
 
-.. autoclass:: BalsamExecutor
+.. autoclass:: libensemble.executors.balsam_executor.BalsamExecutor
   :show-inheritance:
   :members: __init__, register_app, submit_allocation, revoke_allocation, submit
 
-.. autoclass:: BalsamTask
+.. autoclass:: libensemble.executors.balsam_executor.BalsamTask
   :show-inheritance:
   :member-order: bysource
   :members: poll, wait, kill
