+++ docs/executor/executor.rst
@@ -13,7 +13,6 @@ See the Executor APIs for optional arguments.
    :caption: Alternative Executors:
 
    mpi_executor
-   legacy_balsam_executor
    balsam_2_executor
 
 Executor Class
