--- docs/executor/legacy_balsam_executor.rst
@@ -1,17 +1,0 @@
-Legacy Balsam MPI Executor
-==========================
-
-.. automodule:: legacy_balsam_executor
-  :no-undoc-members:
-
-.. autoclass:: LegacyBalsamMPIExecutor
-  :show-inheritance:
-  :inherited-members:
-  :member-order: bysource
-  :members: __init__, submit, poll, manager_poll, kill, set_kill_mode
-
-.. autoclass:: LegacyBalsamTask
-  :show-inheritance:
-  :member-order: bysource
-  :members: workdir_exists, file_exists_in_workdir, read_file_in_workdir, stdout_exists, read_stdout
-  :inherited-members:
