+++ libensemble/executors/__init__.py
@@ -1,4 +1,11 @@
+import importlib
 from libensemble.executors.executor import Executor
 from libensemble.executors.mpi_executor import MPIExecutor
 
-__all__ = ["Executor", "MPIExecutor"]
+if importlib.util.find_spec("balsam"):
+
+    try:
+        from libensemble.executors.balsam_executor import BalsamExecutor
+        __all__ = ["Executor", "MPIExecutor", "BalsamExecutor"]
+    except (ModuleNotFoundError, ImportError):
+        __all__ = ["Executor", "MPIExecutor"]
