+++ libensemble/sim_funcs/executor_hworld.py
@@ -70,7 +70,6 @@ def executor_hworld(H, persis_info, sim_specs, libE_info):
     """Tests launching and polling task and exiting on task finish"""
     exctr = MPIExecutor.executor
     cores = sim_specs["user"]["cores"]
-    USE_BALSAM = "balsam_test" in sim_specs["user"]
     ELAPSED_TIMEOUT = "elapsed_timeout" in sim_specs["user"]
 
     wait = False
@@ -108,18 +107,7 @@ def executor_hworld(H, persis_info, sim_specs, libE_info):
             elif sim_ended_count == 5:
                 args_for_sim = "sleep 2 Fail"  # Manager kill - if signal received else completes
 
-        if USE_BALSAM:
-            task = exctr.submit(
-                calc_type="sim",
-                num_procs=cores,
-                app_args=args_for_sim,
-                hyperthreads=True,
-                machinefile="notused",
-                stdout="notused",
-                wait_on_start=True,
-            )
-        else:
-            task = exctr.submit(calc_type="sim", num_procs=cores, app_args=args_for_sim, hyperthreads=True)
+        task = exctr.submit(calc_type="sim", num_procs=cores, app_args=args_for_sim, hyperthreads=True)
 
         if wait:
             task.wait()
@@ -131,34 +119,12 @@ def executor_hworld(H, persis_info, sim_specs, libE_info):
                 calc_status = TASK_FAILED
 
         else:
-            if sim_ended_count >= 2 and not USE_BALSAM:
+            if sim_ended_count >= 2:
                 calc_status = exctr.polling_loop(task, timeout=timeout, delay=0.3, poll_manager=True)
                 if sim_ended_count == 2 and task.stdout_exists() and "Error" in task.read_stdout():
                     calc_status = WORKER_KILL_ON_ERR
             else:
                 task, calc_status = custom_polling_loop(exctr, task, timeout)
-
-        if USE_BALSAM:
-            task.read_file_in_workdir("ensemble.log")
-            try:
-                task.read_stderr()
-            except ValueError:
-                pass
-
-            task = exctr.submit(
-                app_name="sim_hump_camel_dry_run",
-                num_procs=cores,
-                app_args=args_for_sim,
-                hyperthreads=True,
-                machinefile="notused",
-                stdout="notused",
-                wait_on_start=True,
-                dry_run=True,
-                stage_inout=os.getcwd(),
-            )
-
-            task.poll()
-            task.wait()
 
     else:
         launch_shc = True
