--- libensemble/tests/regression_tests/.bal_coveragerc
@@ -1,32 +1,0 @@
-[run]
-source = $LIBE_PATH
-
-omit =
-    */tests/*
-    */unit_tests/*
-    */unit_tests_nompi/*
-    */regression_tests/*
-    */functionality_tests/*
-    */branin_*/*
-
-branch = true
-
-data_file = .cov_reg_out
-
-parallel = true
-
-[report]
-omit =
-    */__init__.py/*
-    */setup.py/*
-    */tests/*
-    */unit_tests/*
-    */unit_tests_nompi/*
-    */regression_tests/*
-    */functionality_tests/*
-
-exclude_lines =
-    if __name__ == .__main__.:
-
-[html]
-directory = cov_reg
