+++ libensemble/tests/regression_tests/common.py
@@ -63,102 +63,3 @@ def build_borehole():
 
     buildstring = "gcc -o borehole.x ../unit_tests/simdir/borehole.c -lm"
     subprocess.check_call(buildstring.split())
-
-
-def modify_Balsam_worker():
-    # Balsam is meant for HPC systems that commonly distribute jobs across many
-    #   nodes. Due to the nature of testing Balsam on local or CI systems which
-    #   usually only contain a single node, we need to change Balsam's default
-    #   worker setup so multiple workers can be run on a single node.
-    import balsam
-
-    new_lines = [
-        "        for idx in range(10):\n",
-        "            w = Worker(1, host_type='DEFAULT', num_nodes=1)\n",
-        "            self.workers.append(w)\n",
-    ]
-
-    workerfile = "worker.py"
-    balsam_path = os.path.dirname(balsam.__file__) + "/launcher"
-    balsam_worker_path = os.path.join(balsam_path, workerfile)
-
-    with open(balsam_worker_path, "r") as f:
-        lines = f.readlines()
-
-    if lines[-3] != new_lines[0]:
-        lines = lines[:-2]  # effectively inserting new_lines[0] above
-        lines.extend(new_lines)
-
-    with open(balsam_worker_path, "w") as f:
-        for line in lines:
-            f.write(line)
-
-
-def modify_Balsam_pyCoverage():
-    # Tracking line coverage through our tests requires running the Python module
-    #   'coverage' directly. Balsam explicitly configures Python runs with
-    #   'python [script].py args' with no current capability for specifying
-    #   modules. This hack specifies the coverage module and some options.
-    import balsam
-
-    rcfile = os.path.abspath("./libensemble/tests/regression_tests/.bal_coveragerc")
-
-    old_line = "            path = ' '.join((exe, script_path, args))\n"
-    new_line = (
-        "            path = ' '.join((exe, '-m coverage run "
-        + f"--parallel-mode --rcfile={rcfile}', script_path, args))\n"
-    )
-
-    commandfile = "cli_commands.py"
-    balsam_path = os.path.dirname(balsam.__file__) + "/scripts"
-    balsam_commands_path = os.path.join(balsam_path, commandfile)
-
-    with open(balsam_commands_path, "r") as f:
-        lines = f.readlines()
-
-    for i in range(len(lines)):
-        if lines[i] == old_line:
-            lines[i] = new_line
-
-    with open(balsam_commands_path, "w") as f:
-        for line in lines:
-            f.write(line)
-
-
-def modify_Balsam_settings():
-    # Set $HOME/.balsam/settings.json to DEFAULT instead of Theta worker setup
-    settingsfile = os.path.join(os.environ.get("HOME"), ".balsam/settings.json")
-    with open(settingsfile, "r") as f:
-        lines = json.load(f)
-
-    lines["MPI_RUN_TEMPLATE"] = "MPICHCommand"
-    lines["WORKER_DETECTION_TYPE"] = "DEFAULT"
-
-    with open(settingsfile, "w") as f:
-        json.dump(lines, f)
-
-
-def modify_Balsam_JobEnv():
-    # If Balsam detects that the system on which it is running contains the string
-    #   'cc' in its hostname, then it thinks it's on Cooley! If hostnames are
-    #   randomly generated and occasionally may contain that offending string, then this
-    #   modifies Balsam's JobEnvironment class to not check for 'cc'.
-    import balsam
-
-    bad_line = "        'COOLEY' : 'cooley cc'.split()\n"
-    new_line = "        'COOLEY' : 'cooley'.split()\n"
-
-    jobenv_file = "JobEnvironment.py"
-    balsam_path = os.path.dirname(balsam.__file__) + "/service/schedulers"
-    balsam_jobenv_path = os.path.join(balsam_path, jobenv_file)
-
-    with open(balsam_jobenv_path, "r") as f:
-        lines = f.readlines()
-
-    for i in range(len(lines)):
-        if lines[i] == bad_line:
-            lines[i] = new_line
-
-    with open(balsam_jobenv_path, "w") as f:
-        for line in lines:
-            f.write(line)
