--- libensemble/tests/regression_tests/scripts_used_by_reg_tests/cleanup-balsam-test.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-source balsamdeactivate
