+++ libensemble/tests/scaling_tests/forces/balsam_forces/run_libe_forces_balsam.py
@@ -6,7 +6,7 @@ import numpy as np
 from libensemble.libE import libE
 from libensemble.gen_funcs.sampling import uniform_random_sample
 from forces_simf import run_forces_balsam
-from libensemble.executors.balsam_executors import BalsamExecutor
+from libensemble.executors import BalsamExecutor
 from libensemble.tools import parse_args, add_unique_random_streams
 
 from balsam.api import ApplicationDefinition
