+++ libensemble/tests/scaling_tests/forces/forces_adv/run_libe_forces.py
@@ -11,9 +11,6 @@ from libensemble.tools import parse_args, save_libE_output, add_unique_random_st
 from libensemble import logger
 from forces_support import test_libe_stats, test_ensemble_dir, check_log_exception
 
-# Note the Balsam option here is now LegacyBalsam - see balsam_forces for latest.
-USE_BALSAM = False
-
 PERSIS_GEN = False
 
 if PERSIS_GEN:
@@ -36,16 +33,8 @@ if not os.path.isfile(sim_app):
 if is_manager:
     print(f"\nRunning with {nworkers} workers\n")
 
-
-# Create executor and register sim to it.
-if USE_BALSAM:
-    from libensemble.executors.legacy_balsam_executor import LegacyBalsamMPIExecutor
-
-    exctr = LegacyBalsamMPIExecutor()
-else:
-    from libensemble.executors.mpi_executor import MPIExecutor
-
-    exctr = MPIExecutor()
+from libensemble.executors.mpi_executor import MPIExecutor
+exctr = MPIExecutor()
 exctr.register_app(full_path=sim_app, app_name="forces")
 
 # Note: Attributes such as kill_rate are to control forces tests, this would not be a typical parameter.
