+++ libensemble/tests/unit_tests/test_executor_balsam.py
@@ -54,7 +54,7 @@ def teardown_module(module):
 # This would typically be in the user calling script
 def setup_executor():
     """Set up a Balsam Executor with sim app"""
-    from libensemble.executors.balsam_executors import BalsamExecutor
+    from libensemble.executors import BalsamExecutor
 
     exctr = BalsamExecutor()  # noqa F841
 
@@ -89,7 +89,7 @@ def test_submit_app_defaults():
     """Test of submitting an App"""
     print(f"\nTest: {sys._getframe().f_code.co_name}\n")
     exctr = Executor.executor
-    with mock.patch("libensemble.executors.balsam_executors.balsam_executor.Job"):
+    with mock.patch("libensemble.executors.balsam_executor.Job"):
         task = exctr.submit(calc_type="sim")
         task = exctr.submit(app_name="test")
 
@@ -107,7 +107,7 @@ def test_submit_app_workdir():
     """Test of submitting an App with a workdir"""
     print(f"\nTest: {sys._getframe().f_code.co_name}\n")
     exctr = Executor.executor
-    with mock.patch("libensemble.executors.balsam_executors.balsam_executor.Job"):
+    with mock.patch("libensemble.executors.balsam_executor.Job"):
         task = exctr.submit(calc_type="sim", workdir="output", machinefile="nope")
 
     assert task.workdir == os.path.join(exctr.workflow_name, "output"), "workdir not properly defined for new task"
@@ -129,8 +129,8 @@ def test_submit_app_wait():
     """Test of exctr.submit blocking until app is running"""
     print(f"\nTest: {sys._getframe().f_code.co_name}\n")
     exctr = Executor.executor
-    with mock.patch("libensemble.executors.balsam_executors.balsam_executor.Job") as job:
-        with mock.patch("libensemble.executors.balsam_executors.balsam_executor.EventLog") as log:
+    with mock.patch("libensemble.executors.balsam_executor.Job") as job:
+        with mock.patch("libensemble.executors.balsam_executor.EventLog") as log:
             job.return_value.state = "RUNNING"
             log.objects.filter.return_value = [
                 LogEventTest(timestamp=datetime.datetime(2022, 4, 21, 20, 29, 33, 455144))
@@ -148,7 +148,7 @@ def test_submit_revoke_alloc():
     """Test creating and revoking BatchJob objects through the executor"""
     print(f"\nTest: {sys._getframe().f_code.co_name}\n")
     exctr = Executor.executor
-    with mock.patch("libensemble.executors.balsam_executors.balsam_executor.BatchJob"):
+    with mock.patch("libensemble.executors.balsam_executor.BatchJob"):
         alloc = exctr.submit_allocation(site_id="libe-unit-test", num_nodes=1, wall_time_min=30)
 
         assert alloc in exctr.allocations, "batchjob object not appended to executor's list of allocations"
@@ -169,8 +169,8 @@ def test_task_poll():
     """Test of killing (cancelling) a balsam app"""
     print(f"\nTest: {sys._getframe().f_code.co_name}\n")
     exctr = Executor.executor
-    with mock.patch("libensemble.executors.balsam_executors.balsam_executor.Job") as job:
-        with mock.patch("libensemble.executors.balsam_executors.balsam_executor.EventLog"):
+    with mock.patch("libensemble.executors.balsam_executor.Job") as job:
+        with mock.patch("libensemble.executors.balsam_executor.EventLog"):
             task = exctr.submit(calc_type="sim")
 
             job.return_value.state = "PREPROCESSED"
@@ -197,9 +197,9 @@ def test_task_wait():
     """Test of killing (cancelling) a balsam app"""
     print(f"\nTest: {sys._getframe().f_code.co_name}\n")
     exctr = Executor.executor
-    with mock.patch("libensemble.executors.balsam_executors.balsam_executor.Job") as job:
+    with mock.patch("libensemble.executors.balsam_executor.Job") as job:
         with mock.patch(
-            "libensemble.executors.balsam_executors.balsam_executor.EventLog"
+            "libensemble.executors.balsam_executor.EventLog"
         ):  # need to patch since wait polls
             task = exctr.submit(calc_type="sim")
 
@@ -229,10 +229,10 @@ def test_task_kill():
     """Test of killing (cancelling) a balsam app"""
     print(f"\nTest: {sys._getframe().f_code.co_name}\n")
     exctr = Executor.executor
-    with mock.patch("libensemble.executors.balsam_executors.balsam_executor.Job"):
+    with mock.patch("libensemble.executors.balsam_executor.Job"):
         task = exctr.submit(calc_type="sim")
 
-    with mock.patch("libensemble.executors.balsam_executors.balsam_executor.EventLog"):
+    with mock.patch("libensemble.executors.balsam_executor.EventLog"):
         task.kill()
     assert task.finished and task.state == "USER_KILLED", "task not set as killed after kill method"
