--- scripts/balsam/readme.rst
@@ -1,11 +1,0 @@
-=======================
-Balsam analysis scripts
-=======================
-
-These scripts require an activated Balsam database and create plots as ``png`` files.
-
-* ``plot_util_v_time.py``: Shows number of active Balsam launched tasks over time
-
-* ``plot_tasks_v_time.py``: Shows completed Balsam launched tasks over time
-
-* ``plot_waiting_v_time.py``: Shows number of tasks waiting in Balsam database over time
